Update Rise'n Fall - BB bands + RSI  + SMA + Statistics Strategy.xml
Introducing this robot that operates with RSI, measuring buying and selling strength. When one of these forces becomes exhausted in conjunction with Bollinger Bands, a buy or sell order is initiated. The adjustment parameters are straightforward: Stake, Profit, and Loss. Additionally, the entry channels for RSI are customizable, enhancing user flexibility and control over the trading experience.